back_end/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.36.0
dataclasses==0.6
Django==4.2.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
djongo==1.2.31
dnspython==2.4.2
pymongo==4.5.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
